Extract referrals table into a component
