+++ components/Feature/ReferralsTable/index.module.scss
@@ -1,0 +1,8 @@
+.rejected {
+  background-color: red;
+  color: white;
+}
+.approved {
+  background-color: blue;
+  color: white;
+}
